requirements.txt
google-generativeai>=0.3.0 
openai>=1.0.0
